Refactor layout into demo and libraries
