+++ demo/cross_loop.py
@@ -1,9 +1,9 @@
 import time
-from inputs import frame_delay, pulse_button
+from libraries.inputs import frame_delay, pulse_button
 
 def controller_loop(stop_event, controller_states, slot):
     frame = 0
     while not stop_event.is_set():
-        pulse_button(frame, controller_states, slot, triangle=True)
+        pulse_button(frame, controller_states, slot, cross=True)
         frame += 1
         time.sleep(frame_delay)
